# contracts/runs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "运行受理接口交接",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/runs/{run_id}/interruption": {
      "get": {
        "tags": [
          "任务运行"
        ],
        "summary": "Interruption",
        "operationId": "interruption_api_v1_runs__run_id__interruption_get",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/InterruptionView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Interruption Api V1 Runs  Run Id  Interruption Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/resume": {
      "post": {
        "tags": [
          "任务运行"
        ],
        "summary": "Resume",
        "operationId": "resume_api_v1_runs__run_id__resume_post",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs": {
      "post": {
        "tags": [
          "任务运行"
        ],
        "summary": "创建统一运行",
        "description": "sync 最多等待 15 秒；窗口内终结返回 200，否则返回原运行的 202。async/stream 返回 202。相同逻辑请求重发沿用原幂等键，每次均重新授权。",
        "operationId": "admit_api_v1_runs_post",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "已受理，使用原 run_id 查询或订阅",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionReceipt"
                }
              }
            }
          },
          "200": {
            "description": "同步窗口内已终结，须检查 state 和 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "tags": [
          "任务运行"
        ],
        "summary": "Events",
        "operationId": "events_api_v1_runs__run_id__events_get",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "after_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After Sequence"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "最后已处理的事件序号；优先于 after_sequence",
              "title": "Last-Event-Id"
            },
            "description": "最后已处理的事件序号；优先于 after_sequence"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {}
            }
          },
          "410": {
            "description": "事件已过期，查询原运行快照",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "任务运行"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "任务运行"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_v1_runs__run_id__cancel_post",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/v1/runs/{run_id}/interruption": {
      "get": {
        "tags": [
          "运行管理"
        ],
        "summary": "Interruption",
        "operationId": "interruption_admin_v1_runs__run_id__interruption_get",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/InterruptionView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Interruption Admin V1 Runs  Run Id  Interruption Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/v1/runs/{run_id}/resume": {
      "post": {
        "tags": [
          "运行管理"
        ],
        "summary": "Resume",
        "operationId": "resume_admin_v1_runs__run_id__resume_post",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/v1/runs/{run_id}/events": {
      "get": {
        "tags": [
          "运行管理"
        ],
        "summary": "Events",
        "operationId": "events_admin_v1_runs__run_id__events_get",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "after_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After Sequence"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "最后已处理的事件序号；优先于 after_sequence",
              "title": "Last-Event-Id"
            },
            "description": "最后已处理的事件序号；优先于 after_sequence"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {}
            }
          },
          "410": {
            "description": "事件已过期，查询原运行快照",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/v1/runs": {
      "get": {
        "tags": [
          "运行管理"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_admin_v1_runs_get",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "name": "key_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Key Id"
            }
          },
          {
            "name": "purpose",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Purpose"
            }
          },
          {
            "name": "start_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start At"
            }
          },
          {
            "name": "end_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End At"
            }
          },
          {
            "name": "subject_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Type"
            }
          },
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Id"
            }
          },
          {
            "name": "error_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/v1/runs/options": {
      "get": {
        "tags": [
          "运行管理"
        ],
        "summary": "Options",
        "operationId": "options_admin_v1_runs_options_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunFilterOptions"
                }
              }
            }
          }
        },
        "security": [
          {
            "PlatformToken": []
          }
        ]
      }
    },
    "/admin/v1/runs/{run_id}/rerun": {
      "post": {
        "tags": [
          "运行管理"
        ],
        "summary": "Rerun",
        "operationId": "rerun_admin_v1_runs__run_id__rerun_post",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RerunInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/v1/runs/{run_id}/trace": {
      "get": {
        "tags": [
          "运行管理"
        ],
        "summary": "Trace",
        "operationId": "trace_admin_v1_runs__run_id__trace_get",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "after_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After Sequence"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "运行管理"
        ],
        "summary": "Admin Get Run",
        "operationId": "admin_get_run_admin_v1_runs__run_id__get",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/v1/runs/{run_id}/detail": {
      "get": {
        "tags": [
          "运行管理"
        ],
        "summary": "Detail",
        "operationId": "detail_admin_v1_runs__run_id__detail_get",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/v1/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "运行管理"
        ],
        "summary": "Admin Cancel",
        "operationId": "admin_cancel_admin_v1_runs__run_id__cancel_post",
        "security": [
          {
            "PlatformToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "对应响应头 X-Request-ID 的请求标识"
          }
        },
        "type": "object",
        "required": [
          "error",
          "request_id"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterruptionView": {
        "properties": {
          "interruption_id": {
            "type": "string",
            "title": "Interruption Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "state_label": {
            "type": "string",
            "title": "State Label"
          },
          "confirmation_digest": {
            "type": "string",
            "title": "Confirmation Digest"
          },
          "input_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Schema"
          },
          "proposed_input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Proposed Input"
          },
          "can_respond": {
            "type": "boolean",
            "title": "Can Respond"
          },
          "can_approve": {
            "type": "boolean",
            "title": "Can Approve"
          },
          "can_verify": {
            "type": "boolean",
            "title": "Can Verify"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "interruption_id",
          "revision",
          "name",
          "state",
          "state_label",
          "confirmation_digest",
          "input_schema",
          "proposed_input",
          "can_respond",
          "can_approve",
          "can_verify"
        ],
        "title": "InterruptionView"
      },
      "JsonValue": {},
      "RerunInput": {
        "properties": {
          "input": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RerunInput"
      },
...
    },
    "securitySchemes": {
      "PlatformToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "不透明令牌"
      }
    }
  }
}